Improves mouse corner movement (#290)
Refines the corner movement routine for the virtual mouse.

This change ensures correct positioning and prevents unintended behavior
when the virtual mouse is moved to a corner of the screen. It moves
the `bleMouse->move(0, -127)` to the end to be sure that the mouse is
at the top after cornering.

Additionally, only sets the Serial timeout if USB CDC is enabled.