Transaction improvements (#337)
* Refactors device ungrouping process

Simplifies and streamlines the device ungrouping process by
using transactions. This removes the reliance on timers and
message listeners, leading to a more reliable and efficient
operation. Error messages are also improved to provide more
context.

* Adds websocket proxying

Adds support for proxying websocket connections.

This allows the poorxy to handle websocket connections by listening for 'upgrade' events and proxying them to the specified websocket URL.